Refactor ExceptionFieldUtils tests
